change work cover image upload system and resize image in backend

+++ server/src/admin/controllers/work.js
@@ -45,8 +45,9 @@ module.exports = {
             updated: Date.now(),
         };
 
-        console.log(req.file)
-
+        if (req.file) {
+            work.img = req.file.filename
+        }
         let oldWork = await Work.findOne({_id: candidate});
         try {
             const old = await Work.findByIdAndUpdate(
@@ -59,7 +60,9 @@ module.exports = {
             });
 
             // console.log(oldWork)
-            fs.unlinkSync(__dirname + `/../../../_uploads/work/${oldWork.img}`)
+            if (oldWork.img && req.file) {
+                fs.unlinkSync(__dirname + `/../../../_uploads/work/${oldWork.img}`)
+            }
 
             res.status(201).json(work)
         } catch (e) {
